refactor(cli): extract WorkflowOptimizeReportRenderer
